Added DirExists checking. Added console logging

+++ Program.cs
@@ -1,12 +1,19 @@
-﻿string path = @"c:\Users\Admin\Downloads\30da91a108\";
+﻿string path = @"c:\Users\Admin\Downloads\30da91a108";
 
 DirectoryInfo di = new DirectoryInfo(path);
 
-foreach (FileInfo file in di.GetFiles().Where(x => (DateTime.Now - x.LastWriteTime).TotalMinutes > 30))
+if (di.Exists)
 {
-    file.Delete();
+    foreach (FileInfo file in di.GetFiles().Where(x => (DateTime.Now - x.LastWriteTime).TotalMinutes > 1))
+    {
+        file.Delete();
+    }
+    foreach (DirectoryInfo dir in di.GetDirectories().Where(x => (DateTime.Now - x.LastWriteTime).TotalMinutes > 1))
+    {
+        dir.Delete(true);
+    }
 }
-foreach (DirectoryInfo dir in di.GetDirectories().Where(x => (DateTime.Now - x.LastWriteTime).TotalMinutes > 30))
-{
-    dir.Delete(true);
-}
+else
+    Console.WriteLine("Некорректный путь!");
+
+
